+++ src/collections/dic.php
@@ -3,13 +3,10 @@
 namespace Vosiz\Utils\Collections;
 
 require_once(__DIR__.'/../keyval.php');
-require_once(__DIR__.'/../object.php');
-require_once(__DIR__.'/exc.php');
 
+class Dictionary extends Collection {
 
-class Dictionary extends \SmartObject implements \IteratorAggregate {
-
-    private $Temp = []; // temp data, original key/value
+    protected $ToStringName = "Vosiz/Utils/Dictionary";
     private $Kvps = []; // more modern, keyvaluepair-based
 
     /**
@@ -22,47 +19,10 @@ class Dictionary extends \SmartObject implements \IteratorAggregate {
         return new Dictionary($a);
     }
 
-    /** Base constructor - optional values 
-     * @param array $value values to add
-    */
-    public function __construct(array $values = array()) {
-
-        parent::__construct();
-        foreach($values as $k => $v) {
-
-            $this->Add($v, $k);
-        }
-    }
-
     /**
-     * Get property-like magic
-     * @param string $key key/property
-     * @return mixed|null null when not found
+     * Returns iterator
+     * @return KeyValuePairStrict[]
      */
-    public function __get(string $key) {
-
-        $key = str_camel($key);
-        return getifset($this->Temp, $key);
-    }
-
-    /** Classic toString override
-     * @return string toString representation
-    */
-    public function __toString() {
-
-        $text = "Vosiz/Utils/Dictionary";
-        if(!$this->IsEmpty()) {
-            $text .= ": cnt=".$this->Count();
-            $text .= " {";
-            foreach($this->Temp as $key => $item) {
-                $text .= "<br> - [$key] => ".tostr($item);
-            }
-            $text .= " }";
-        }
-
-        return $text;
-    }
-
     public function getIterator(): \Traversable
     {
         foreach ($this->Kvps as $pair) {
@@ -117,7 +77,7 @@ class Dictionary extends \SmartObject implements \IteratorAggregate {
 
                 } else {
 
-                    throw new DictionaryException("Key already exists");
+                    throw new CollectionException("Key already exists");
                 }
 
             } else {
@@ -129,23 +89,6 @@ class Dictionary extends \SmartObject implements \IteratorAggregate {
         }
     }
 
-    /** 
-     * Adds multiple values
-     * @param mixed[] $data objects to add
-     * @param bool $update_existing if true updates value, throws exception on existing key if false
-     * @return bool success
-    */
-    public function AddRange(array $data = array(), bool $update_existing = false) {
-
-        $res = true;
-        foreach($data as $k => $v) {
-
-            if(!$this->Add($v, $k, $update_existing)) 
-                $res = false;
-        }
-
-        return $res;
-    }
 
     /**
      * Remove member(s) from Dictionary
@@ -154,8 +97,8 @@ class Dictionary extends \SmartObject implements \IteratorAggregate {
      */
     public function Remove($index, bool $keep = false) {
 
-        asarray($index);
-        unsetra($this->Temp, $index, $keep);
+        parent::Remove($index, $keep);
+        unsetra($this->Kvps, $index, $keep);
     }
 
     /**
@@ -175,108 +118,8 @@ class Dictionary extends \SmartObject implements \IteratorAggregate {
     */
     public function Clear() {
 
-        $this->Temp = array();
-    }
-
-    /**
-     * Intersect, value-based, distincts values with diff. keys
-     * @param array $a intersection array
-     * @return Dictionary
-     */
-    public function Intersect(array $a = array()) {
-
-        return array_intersect($this->ToArray(), $a);
-    }
-
-    /**
-     * Merge Dictionary to this
-     * @param Dictionary $c Dictionary to merge
-     * @param bool $rewrite when keys are same (true = rewrites with new value)
-     */
-    public function Merge(Dictionary $c, bool $rewrite = false) {
-
-        $this->AddRange($c->ToArray(), $rewrite);
-    }
-
-    /**
-     * Count of members
-     * @return int
-     */
-    public function Count() {
-
-        return count($this->Temp);
-    }
-
-    /**
-     * Has key
-     * @param int|string $key questioned key
-     * @return bool true if present
-     */
-    public function HasKey($key) {
-
-        return array_key_exists($key, $this->Temp);
-    }
-
-    /**
-     * Has value
-     * @param mixed $object questioned value
-     * @return bool true if present
-     */
-    public function HasValue($value) {
-
-        return in_array($value, $this->Temp);
-    }
-
-    /**
-     * Gets index/key by value
-     * @param mixed $value questioned value
-     * @return int|string|false index or fail
-     */
-    public function IndexOf($value) {
-
-        return array_search($value, $this->Temp);
-    }
-
-    /**
-     * Check if it is empty
-     * @return bool empty
-     */
-    public function IsEmpty() {
-
-        return $this->Count() == 0; 
-    }
-
-    /** 
-     * Get keys (all or with filtration)
-     * @param array ...$filter_values every value to filtrate on (none means all keys)
-     * @return mixed keys
-    */
-    public function Keys(...$filter_values): array {
-
-        if (empty($filter_values)) {
-            return array_keys($this->Temp);
-        }
-
-        $keys = [];
-
-        // fixing va as array
-        if(is_array($filter_values))
-            $filter_values = current($filter_values);
-
-        foreach ($filter_values as $value) {
-            $keys = array_merge($keys, array_keys($this->Temp, $value, true));
-        }
-
-        return array_unique($keys);
-    }
-
-    /**
-     * Convert to array
-     * @return array
-     */
-    public function ToArray() {
-
-        return $this->Temp;
+        parent::Clear();
+        $this->Kvps = array();
     }
 
     /**
@@ -288,60 +131,4 @@ class Dictionary extends \SmartObject implements \IteratorAggregate {
         return $this->Kvps;
     }
 
-    /**
-     * Get values
-     * @return array
-     */
-    public function Values() {
-
-        return array_values($this->Temp);
-    }
-
-    /**
-     * Alias for ToArray
-     * @return array
-     */
-    public function AsArray() {
-        
-        return $this->Temp;
-    }
-
-    /**
-     * Alias for Count
-     * @return int
-     */
-    public function Length() {
-
-        return $this->Count();
-    }
-
-    /**
-     * Alias for HasValue
-     * @param mixed $object questioned value
-     * @return bool true if present
-     */
-    public function Contains($object) {
-
-        return $this->HasValue($object);
-    }
-
-    /**
-     * Alias to IndexOf
-     * @param mixed $value questioned value
-     * @return int|string|false index
-     */
-    public function Find($value) {
-
-        return $this->IndexOf($value);
-    }
-
-    /**
-     * Finds all indeces of values
-     * @param array $values Searched
-     * @return array
-     */
-    public function FindAll($values = array()) {
-
-        return $this->Keys($values);
-    }
 }